requirements.txt
tqdm
numpy
Pillow
matplotlib
tensorboardX
